backend/cloud_functions/rag_lro_monitor/requirements.txt
functions-framework==3.*
google-cloud-aiplatform>=1.47.0
google-api-core>=2.0.0
supabase>=1.0.0
# Add any other specific dependencies
